requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.1
sqlalchemy==2.0.31
asyncpg==0.31.0
geoalchemy2==0.15.0
redis==5.0.4
httpx==0.27.0
python-dotenv==1.0.1
pydantic[email]==2.8.2
python-multipart==0.0.9
aiosqlite==0.19.0  # для локальных тестов, если нет PostGIS
pydantic-settings==2.12.0